[REF] stock, stock_delivery: move delviery data to inventory
The field are usefull in the incoming dispatching sytem. So it would be
a good idea to share them
