+++ addons/stock_delivery/models/stock_quant_package.py
@@ -50,4 +50,3 @@ class StockQuantPackage(models.Model):
     weight_uom_name = fields.Char(string='Weight unit of measure label', compute='_compute_weight_uom_name', readonly=True, default=_get_default_weight_uom)
     weight_is_kg = fields.Boolean("Technical field indicating whether weight uom is kg or not (i.e. lb)", compute="_compute_weight_is_kg")
     weight_uom_rounding = fields.Float("Technical field indicating weight's number of decimal places", compute="_compute_weight_is_kg")
-    shipping_weight = fields.Float(string='Shipping Weight', help="Total weight of the package.")
